stert to add redux to the scanner
