+++ src/features/userData/userDataSlice.js
@@ -19,7 +19,17 @@ export const weightSlice = createSlice({
     fetchWeightFulfiled: (state, action) => ({ ...state, ...action.payload }),
   },
 });
+
+export const barcodeDataSlice = createSlice({
+  name: 'barcode',
+  initialState: null,
+  reducers: {
+    setBarcodeData: (state, action) => action.payload,
+  },
+});
 export const { fetchUserDataFulfiled } = userDataSlice.actions;
 export const { fetchWeightFulfiled } = weightSlice.actions;
+export const { setBarcodeData } = barcodeDataSlice.actions;
 export const userReducer = userDataSlice.reducer;
 export const weighReducer = weightSlice.reducer;
+export const barcodeReducer = barcodeDataSlice.reducer;
