+++ src/store/store.js
@@ -2,7 +2,11 @@ import { configureStore } from '@reduxjs/toolkit';
 import dateSlice from '../features/date/dateSlice';
 import loadingSlice from '../features/loading/loadingSlice';
 import mealSlice from '../features/meals/mealSlice';
-import { userReducer, weighReducer } from '../features/userData/userDataSlice';
+import {
+  barcodeReducer,
+  userReducer,
+  weighReducer,
+} from '../features/userData/userDataSlice';
 
 export const store = configureStore({
   reducer: {
@@ -11,5 +15,6 @@ export const store = configureStore({
     loading: loadingSlice,
     userData: userReducer,
     weight: weighReducer,
+    barcode: barcodeReducer,
   },
 });
